+++ .gitignore
@@ -1,3 +1,3 @@
 db.sqlite3
 .idea/
-conapp_checker/__pycache__/
+app/__pycache__/
