+++ app/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for conapp_checker project.
+Django settings for app project.
 
 Generated by 'django-admin startproject' using Django 1.11.7.
 
@@ -47,10 +47,10 @@ MIDDLEWARE = [
     'django.contrib.auth.middleware.AuthenticationMiddleware',
     'django.contrib.messages.middleware.MessageMiddleware',
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
-    'conapp_checker.middleware.CustomSocialAuthExceptionMiddleware',
+    'app.middleware.CustomSocialAuthExceptionMiddleware',
 ]
 
-ROOT_URLCONF = 'conapp_checker.urls'
+ROOT_URLCONF = 'app.urls'
 
 TEMPLATES = [
     {
@@ -71,7 +71,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'conapp_checker.wsgi.application'
+WSGI_APPLICATION = 'app.wsgi.application'
 
 # Database
 # https://docs.djangoproject.com/en/1.11/ref/settings/#databases
